+++ MyDubDubGrub/Utilities/Managers/LocationManager.swift
@@ -9,4 +9,6 @@ import Foundation
 
 final class LocationManager: ObservableObject {
 	@Published var locations: [DDGLocation] = []
+	
+	var selectedLocation: DDGLocation?
 }
