+++ MyDubDubGrub/Views/Screens/LocationMapView/LocationMapView.swift
@@ -22,7 +22,13 @@ struct LocationMapView: View {
     var body: some View {
 		ZStack {
 			Map(coordinateRegion: $viewModel.region, showsUserLocation: true, annotationItems: locationManager.locations) { location in
-				MapMarker(coordinate: location.location.coordinate, tint: .brandPrimary)
+				MapAnnotation(coordinate: location.location.coordinate, anchorPoint: CGPoint(x: 0.5, y: 0.75)) {
+					DDGAnnotation(location: location)
+						.onTapGesture {
+							locationManager.selectedLocation = location
+							viewModel.isShowingDetailView = true
+						}
+				}
 			}
 			.accentColor(Color.red)
 			.ignoresSafeArea()
@@ -33,15 +39,20 @@ struct LocationMapView: View {
 				Spacer()
 			}
 		}
-		.sheet(isPresented: $viewModel.isShowingOnboardView, onDismiss: viewModel.checkIfLocationServicesIsEnabled) {
-			OnBoardView(isShowingOnboardView: $viewModel.isShowingOnboardView)
+		.sheet(isPresented: $viewModel.isShowingDetailView) {
+			NavigationView {
+				LocationDetailView(viewModel: LocationDetailViewModel(location: locationManager.selectedLocation!))
+					.toolbar {
+						Button(action: { viewModel.isShowingDetailView = false }, label: {
+							XDismissButton()
+						})
+					}
+			}
 		}
 		.alert(item: $viewModel.alertItem, content: { alertItem in
 			Alert(title: alertItem.title, message: alertItem.message, dismissButton: alertItem.dismissButton)
 		})
-		.onAppear {
-			viewModel.runStartUpChecks()
-			
+		.onAppear {			
 			if locationManager.locations.isEmpty {
 				viewModel.getLocations(for: locationManager)
 			}
@@ -49,6 +60,6 @@ struct LocationMapView: View {
     }
 }
 
-#Preview {
-    LocationMapView()
-}
+//#Preview {
+//    LocationMapView()
+//}
