+++ MyDubDubGrub/Views/Screens/LocationMapView/LocationMapViewModel.swift
@@ -6,11 +6,9 @@
 //
 
 import MapKit
-import CoreLocation
 
-final class LocationMapViewModel: NSObject, ObservableObject {
+final class LocationMapViewModel: ObservableObject {
 	// MARK: - Properties
-	@Published var isShowingOnboardView = false
 	@Published var alertItem: AlertItem?
 	@Published var region = MKCoordinateRegion(
 		center: CLLocationCoordinate2D(
@@ -22,50 +20,8 @@ final class LocationMapViewModel: NSObject, ObservableObject {
 			longitudeDelta: 0.01
 		)
 	)
-	
-	var deviceLocationManager: CLLocationManager?
-	let kHasSeenOnboardView = "hasSeenOnboardView"
-	
-	var hasSeenOnboardView: Bool {
-		return UserDefaults.standard.bool(forKey: kHasSeenOnboardView)
-	}
-	
-	// MARK: - Functions
-	func runStartUpChecks() {
-		if !hasSeenOnboardView {
-			isShowingOnboardView = true
-			UserDefaults.standard.set(true, forKey: kHasSeenOnboardView)
-		} else {
-			checkIfLocationServicesIsEnabled()
-		}
-	}
-	
-	func checkIfLocationServicesIsEnabled() {
-		if CLLocationManager.locationServicesEnabled() {
-			self.deviceLocationManager = CLLocationManager()
-			self.deviceLocationManager!.delegate = self
-		} else {
-			self.alertItem = AlertContext.locationDisabled
-		}
-	}
-	
-	private func checkLocationAuthorization() {
-		guard let deviceLocationManager = deviceLocationManager else { return }
+	@Published var isShowingDetailView = false
 		
-		switch deviceLocationManager.authorizationStatus {
-		case .notDetermined:
-			deviceLocationManager.requestWhenInUseAuthorization()
-		case .restricted:
-			alertItem = AlertContext.locationRestricted
-		case .denied:
-			alertItem = AlertContext.locationDenied
-		case .authorizedAlways, .authorizedWhenInUse:
-			break
-		@unknown default:
-			break
-		}
-	}
-	
 	func getLocations(for locationManager: LocationManager) {
 		CloudKitManager.shared.getLocations { result in
 			DispatchQueue.main.async {
@@ -77,11 +33,5 @@ final class LocationMapViewModel: NSObject, ObservableObject {
 				}
 			}
 		}
-	}
-}
-
-extension LocationMapViewModel: CLLocationManagerDelegate {
-	func locationManagerDidChangeAuthorization(_ manager: CLLocationManager) {
-		checkLocationAuthorization()
 	}
 }
